Exeptions
Sometimes the f1calendar page defaults to displaying the upcoming race as a race that has already taken place (the content of the page is updated dynamically when you access it), so that the data are downloaded for the race weekend that has already taken place and the remaining time until the race is calculated, which always has a negative value. Exeption is for that case to prevent stopping program execution.